shorten role session names
